+++ MCMA/app/src/main/java/vn/edu/usth/mcma/frontend/ConnectAPI/Retrofit/APIs/BookingProcessAPIs/GetDateScheduleByScreenIdAPI.java
@@ -1,7 +1,5 @@
 package vn.edu.usth.mcma.frontend.ConnectAPI.Retrofit.APIs.BookingProcessAPIs;
 
-import java.util.List;
-
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
@@ -10,7 +8,7 @@ import vn.edu.usth.mcma.frontend.ConnectAPI.Model.Response.BookingProcess.Schedu
 public interface GetDateScheduleByScreenIdAPI {
     @GET("/api/v1/user/booking/allSchedulesByMovieAndCinemaAndScreen")
     Call<ScheduleResponse> getAllSchedulesByMovieAndCinemaAndScreen(
-            @Query("movieId") Integer movieId,
+            @Query("movieId") Long movieId,
             @Query("cinemaId") Integer cinemaId,
             @Query("screenId") Integer screenId
     );
